Use proper `Id` in realm API

+++ scripts/schema.sql
@@ -1,6 +1,6 @@
 create table realms (
-    id int generated always as identity (start with 0 minvalue 0) primary key,
-    parent int not null references realms on delete restrict,
+    id bigint generated always as identity (start with 0 minvalue 0) primary key,
+    parent bigint not null references realms on delete restrict,
     name text not null
 );
 -- To truncate this and restart the primary key counter:
